requirements.txt
github-copilot-sdk>=0.1.2
python-dotenv>=1.0.0
